Do not require a specific profile section in the athenacli config
We use some kind of mechanism to generate temporary credentials. These show up in $HOME/.aws/credentials
with an additional aws_session_token in a separate profile. The change here makes it possible to use these credentials
without adding a specific section for this temporary profile to the athena config file.

Before this change this would fail as the profile would not be found and error out.