Allow starting without a athenaclirc file
If you start now with `athenacli --athenaclirc=` the default config file is not generated and no error is
thrown. You of course have to set all config on the commandline.

This allows us to use athenacli in a docker container without generating a config file or run it twice to get arround
the intial config file generation and sys.exit(1).