Do not require a specific profile section in the athenacli config and allow not to generate the default config file (#54)
* Do not require a specific profile section in the athenacli config

We use some kind of mechanism to generate temporary credentials. These show up in $HOME/.aws/credentials
with an additional aws_session_token in a separate profile. The change here makes it possible to use these credentials
without adding a specific section for this temporary profile to the athena config file.

Before this change this would fail as the profile would not be found and error out.

* Allow starting without a athenaclirc file

If you start now with `athenacli --athenaclirc=` the default config file is not generated and no error is
thrown. You of course have to set all config on the command line.

This allows us to use athenacli in a docker container without generating a config file or run it twice to get around
the initial config file generation and sys.exit(1).